Improve dashboard + footer
